requirements_test.txt
# Development and Testing Requirements
pytest
pytest-asyncio
pytest-homeassistant-custom-component
icalendar>=5.0.0
